# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "AssetRef": {
        "properties": {
          "index": {
            "title": "Index",
            "type": "integer"
          },
          "mime": {
            "title": "Mime",
            "type": "string"
          },
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "index",
          "url",
          "mime"
        ],
        "title": "AssetRef",
        "type": "object"
      },
      "CandidateType": {
        "properties": {
          "confidence": {
            "title": "Confidence",
            "type": "number"
          },
          "jewelry_type": {
            "$ref": "#/components/schemas/JewelryType"
          }
        },
        "required": [
          "jewelry_type",
          "confidence"
        ],
        "title": "CandidateType",
        "type": "object"
      },
      "ErrorDetail": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorDetail",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JewelryType": {
        "description": "Reconciled against the client's real matrix (Sheet1 header row) in Phase 0a\nStep 6. The client's sheet is the source of truth (docs/business-rules.md R9).",
        "enum": [
          "RING",
          "NECKLACE",
          "EARRING",
          "BRACELET",
          "BANGLE",
          "ANKLET",
          "HIPBELT"
        ],
        "title": "JewelryType",
        "type": "string"
      },
      "JobListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobResponse"
            },
            "title": "Jobs",
            "type": "array"
          }
        },
        "required": [
          "jobs",
          "count"
        ],
        "title": "JobListResponse",
        "type": "object"
      },
      "JobResponse": {
        "description": "Mirrors the GET /jobs/{job_id} example in docs/api-routes.md.\n\n`jewelry_type` maps to `Job.jewelry_type_final`, falling back to\n`Job.jewelry_type_requested` when the final type hasn't been set yet\n(e.g. while still `queued`/`classifying`) \u2014 this is a read-model\nconvenience so callers see the client-provided type immediately rather\nthan `null` until classification completes. The route that builds this\nfrom a `Job` (Step 4) is responsible for applying that fallback; this\nschema only declares the shape.",
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetRef"
            },
            "title": "Assets",
            "type": "array"
          },
          "candidate_types": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CandidateType"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Candidate Types"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "deadline_at": {
            "title": "Deadline At",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ErrorDetail"
              },
              {
                "type": "null"
              }
            ]
          },
          "jewelry_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JewelryType"
              },
              {
                "type": "null"
              }
            ]
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "mock": {
            "title": "Mock",
            "type": "boolean"
          },
          "service": {
            "$ref": "#/components/schemas/ServiceType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "type_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Type Source"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status",
          "service",
          "jewelry_type",
          "type_source",
          "confidence",
          "mock",
          "created_at",
          "updated_at",
          "completed_at",
          "deadline_at",
          "assets",
          "candidate_types",
          "error"
        ],
        "title": "JobResponse",
        "type": "object"
      },
      "JobStatus": {
        "enum": [
          "queued",
          "classifying",
          "resolving",
          "submitting",
          "generating",
          "storing",
          "succeeded",
          "failed",
          "needs_input",
          "needs_review"
        ],
        "title": "JobStatus",
        "type": "string"
      },
...
      "ResolveRequest": {
        "properties": {
          "jewelry_type": {
            "$ref": "#/components/schemas/JewelryType"
          }
        },
        "required": [
          "jewelry_type"
        ],
        "title": "ResolveRequest",
        "type": "object"
      },
      "ServiceType": {
        "description": "v1 values reconciled against the client's real matrix in Phase 0a Step 6: the\nsheet crosses 4 model/context categories with 2 styles (Traditional/Modern),\nnot the originally-assumed MODEL_SHOT/CATALOG_WHITE/LIFESTYLE/CLOSEUP_MACRO.",
        "enum": [
          "FEMALE_MODEL_TRADITIONAL",
          "FEMALE_MODEL_MODERN",
          "MALE_MODEL_TRADITIONAL",
          "MALE_MODEL_MODERN",
          "MANNEQUIN_TRADITIONAL",
          "MANNEQUIN_MODERN",
          "PRODUCT_STYLING_TRADITIONAL",
          "PRODUCT_STYLING_MODERN",
          "REMOVE_BG",
          "CHANGE_BG",
          "MIX_PIECES"
        ],
        "title": "ServiceType",
        "type": "string"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Jewellery Generation Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "operationId": "list_jobs_route_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/JobStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs Route"
      }
    },
    "/api/v1/jobs/{job_id}": {
      "get": {
        "operationId": "get_job_route_api_v1_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Route"
      }
    },
    "/api/v1/jobs/{job_id}/assets/{index}": {
      "get": {
        "operationId": "get_job_asset_route_api_v1_jobs__job_id__assets__index__get",
        "parameters": [
          {
            "in": "path",
            "name": "index",
            "required": true,
            "schema": {
              "title": "Index",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job Asset Route"
      }
    },
    "/api/v1/jobs/{job_id}/resolve": {
      "post": {
        "operationId": "resolve_job_route_api_v1_jobs__job_id__resolve_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve Job Route"
      }
    },
...
  }
}